[JAY-591] Add Github action for releasing the gem
This Github action runs whenever a Pull Request that changes the
lib/jay_api/version.rb file is merged.

It builds the gem, creates a release (with a corresponding tag) and
pushes the generated gem package to rubygems.org. It uses Rubygems's
trusted publishers system hence there is no need for an OTP code.