Add missing Elasticsearch modules and requires
Without these in place some of the Elasticsearch related classes did not
get loaded when the console was invoked.

Having these in place also ensures an accurate code-coverage report
since the files will be loaded even if they are never explicitly loaded
by the tests, hence showing up as not covered in the report.