[JAY-495] Allow false, Array and Hash in #source
Change the QueryBuilder#source method to allow filter_expr to take
false, arrays and hashes in addition to simple strings.

The main driver behind this change is allowing the use of false to
remove the document's source from the result JSON. Array and Hash are
being added in order to match the method's implementation to
Elasticsearch's documentation.