[JAY-623] Allow path to be omitted in GitilesHelper#gitiles_url
Changes the method to allow the path parameter to be omitted (or given
as nil). In such cases the method generates a link to the refspec
itself.

This is needed because without this change there is no way to generate
links to the refspec itself. Even if nil were to be given as value for
the path parameter the generated link would have a / at the end, which
is a link to the root directory, not the refspec.

The difference is, that a link to the refspec adds the commit
information (commit message, parent, committer, etc) to the top of the
page.