[JAY-409] Add #ping to Elasticsearch::Client
The method allows the user to ping the Elasticsearch cluster to check
connectivity and latency, without the need for any previous knowledge
about the cluster, like endpoint or index names.

The method just forwards the call to the underlying Elasticsearch
Transport Client's #ping method. Which returns true when the ping is
successful and false when it fails. Although, less common, it can also
raise a Transport::Errors::Forbidden when the user doesn't have
permissions to ping the cluster.