[JAY-653] Add Configuration#keys
The method allows the caller to get the list of keys that the
configuration object has. This could be used to iterate over the
elements in the Configuration object or to validate the object's
structure without having to turn it into a Hash first (#to_h).

Note that the addition of this method means that it is no longer
possible to access the value of a an attribute called 'keys' via the
dot syntax, however, it is still possible to access its value using the
brackets: []