[JAY-655] Add the Elasticsearch::Stats::Index class
The class holds information about an Elasticsearch index. It is meant
to be used by the Elasticsearch::Stats class, which for the moment is
just an stub.

In an upcoming commit the class will be filled with the needed logic to
retrieve information about the indices in the Elasticsearch cluster.