[JAY-655] Add the Elasticsearch::Stats::Indices class
The class holds information about the Indices Statistics returned by
Elasticsearch's Stats API. The class has a little bit of logic to
separate system and user-defined indices. It also relies on lazy
enumerators to avoid instantiating too many objects when they aren't
needed.

In an upcoming commit the class will be used by the Elasticsearch::Stats
class, which at the time of writing it's only a stub.