[JAY-656] Add the Elasticsearch::Stats::Node class
The class holds information about one of the nodes that make up the
Elasticsearch cluster. For the moment it only provides the #storage
method, which gives the caller access to storage-related information
about the node.

The class will be used in an upcoming commit to give the user access to
said data through the Elasticsearch::Client class.