[JAY-656] Add Elasticsearch::Stats::Nodes class
The class holds the information for each of the nodes that make up the
Elasticsearch cluster. It acts as a collection of Node objects, one for
each node.

The class uses a lazy enumerator to avoid instantiating Node objects
needlessly, they are only initialized when they are accessed.

The class will be used in an upcoming commit to give the user access to
node-related statistics through the Elasticsearch::Client class.