[JAY-625] Add the Aggregations::Sources class
This class holds a collection of sources for a Composite aggregation for
Elasticsearch.

The class will be used in an upcoming commit by said aggregation to hold
its sources and to allow the user to add sources to a composite
aggregation by passing a block to its constructor.

For the moment the class can only handle a Terms source. Additional
sources might be added in the future.