[JAY-676] Add the Elasticsearch::Indexes class
The class represents a set of Elasticsearch indexes. Through the class
it is possible to query multiple indexes or push data to them at the
same time.

The class doesn't really add any logic. All the functional logic is just
taken from the Indexable mixin.

Because of the way the #push method works. When the batch_size is not a
multiple of the number of indexes it is possible for the pushed batch
to be slightly bigger than batch_size. Hence a Warning is logged when
this condition is detected by the Indexes class's constructor.