[JAY-676] Move code from Index to Indexable
Creates the Indexable mixin and moves much of the code from the
Elasticsearch::Index class there. These methods can actually be used
with a single index or with multiple indexes.

In an upcoming commit these methods will be used by another class which
will allow multiple indexes to be searched. Data could also be pushed to
multiple indexes at the same time.