[JAY-676] Move common test code to a shared file
Moves the test code for the methods that were moved to the Indexable
mixin to a shared examples file. This test code will be used in an
upcoming commit in order to test other classes that include the mixin.

Note: Some changes were made to the test code in order to allow it to
work in a shared_examples / shared_context block. These changes are
mainly inconsequential, for example, moving the parameters of a method
call to a separate let.