[JAY-676] Modify Indexable to work with multiple indexes
This is possible because some of the methods of the underlying
Elasticsearch client can accept and act over multiple indexes.

For those that can only act over a single index, like #index and #bulk
extra logic is added to handle this on the Jay API side.

In order to keep backwards compatibility a small amount of logic is
added to the Index class so that the constructor parameters and the
return values of the methods can stay the same.