[ELITERT-1275] Increase Ruby dependency to 3.1.0
This is needed in order to be able to update the version of the git gem
to '~> 3'. Starting from 3.0.0 the git gem requires Ruby 3.0.0 or later.
Furthermore, it depends on process_executer '~> 1.3', which requires
Ruby 3.1.0 or later.

Version 3.0.0 of git includes a fix that prevents a crash when the
method #branch is called on a repository currently in a detached head.