[JAY-678] Allow nested Elasticsearch boolean queries
Changes the QueryBuilder::QueryClauses::Bool class to allow it to accept
nested boolean queries.

The main driver behind this change is the need to combine 'must' and
'should' clauses with a particular precedence when querying multiple
indexes, especially when some of the fields in the query exist in only a
subset of those indexes.