[JAY-706] Add #clone to QueryBuilder
Adds a #clone method to the QueryBuilder class. The method properly
clones the QueryBuilder instance and its nested objects.

Without this method when #clone was called on a QueryBuilder object,
changes to the clone's query, aggregations, sort and, in certain
cases, source clauses would propagate to the original object.

With the addition of the #clone method these issues no longer happen.