[JAY-718] Add 'timeout' parameter to ClientFactory
Adds the 'timeout' parameter to Elasticsearch::ClientFactory#create. The
parameter allows the user to specify the number of seconds the
underlying HTTP client will wait for Elasticsearch's response.

The main driver behind this change is to alleviate some timeout errors
that have been happening for complicated or large queries fetching big
amounts of data.