[BUGFIX] Revert 9291f29 "[JAY-639] Fix PropertiesFetcher#last"
The commit being reverted was supposed to fix an issue in the
PropertiesFetcher class's #last method, however, that issue never
existed to begin with.

The argument for the fix was that #last was sorting in descending order
and then taking the last, which would effectively return the first,
however, that is not true, because it is also asking Elasticsearch for
a single document (by calling #size with its parameter set to 1).

Hence, Elasticsearch would return a single document, the last entry in
the index (ordered chronologically). So the logic was correct to begin
with.

This fixes the issue by reverting the changes introduced by #11